Added a little more accuracy on completions...
I don't think this is done, but I don't see any inconsistency to clean up
regarding how to react to the FLAG argument in a programmed completions
function.  I added an annotation function for example, but no behavior was
observed.  Most users likely don't want to go this far anyway.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>